LibAudio: fix some compile errors for android P [26/28]
PD# NONE

p compile system is very strict, many warnings become errors
add LOCAL_CFLAGS still can not be resolved. must fix it

+++ audio_codec/libamr/dec_main.c
@@ -686,7 +686,7 @@ Word32 D_MAIN_decode(Word16 mode, Word16 prms[], Word16 synth16k[],
 
             if ((L_tmp > MIN_16) & (L_tmp < 32768)) {
                 exc[i + i_subfr] = (Word16)L_tmp;
-                tmp = (Word16)(abs(L_tmp));
+                tmp = (Word16)(abs((int)L_tmp));
 
                 if (tmp > max) {
                     max = tmp;
